blacked all files; updated package structure; fixed PlainNet for training in dp & ddp mode
